Reorganize the structure of the Search
