+++ src/Components/Header/PageHeader.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Row, Col, Avatar, Layout } from 'antd';
 import { AlignCenterOutlined, UserOutlined } from '@ant-design/icons';
 
-import { Search } from '../../Components';
+import { Search } from '../../Search';
 
 import styles from './PageHeader.module.scss';
 
@@ -23,15 +23,22 @@ const PageHeader: React.FC<Props> = (props) => {
         <Col>
           <Row type="flex">
             <Col>
-              <AlignCenterOutlined style={{ cursor: 'pointer', marginLeft: '-32px' }} onClick={props.openSideBar} />
+              <AlignCenterOutlined
+                style={{ cursor: 'pointer', marginLeft: '-32px' }}
+                onClick={props.openSideBar} />
             </Col>
             <Col>
               <Search />
             </Col>
           </Row>
         </Col>
-        <Col className={styles.avatar} style={{ marginRight: !collapsed && !isMobile ? '12rem' : '' }}>
-          <Avatar icon={<UserOutlined />} className={styles.icon} />
+        <Col
+          className={styles.avatar}
+          style={{ marginRight: !collapsed && !isMobile ? '12rem' : '' }}>
+          <Avatar
+            icon={<UserOutlined />}
+            className={styles.icon}
+          />
         </Col>
       </Row>
     </Header>
