+++ src/Components/index.ts
@@ -2,6 +2,5 @@ import { Header, PageHeader } from './Header';
 import { Button } from './Button';
 import { Input, PasswordInput } from './Input';
 import { Icon } from './Icon';
-import { Search } from './Search';
 
-export { Header, Button, Input, PasswordInput, PageHeader, Icon, Search };
+export { Header, Button, Input, PasswordInput, PageHeader, Icon };
